Expense_backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
bcrypt==3.2.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
google-generativeai==0.8.4
requests==2.31.0
Werkzeug==3.1.3

